+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+# [1.1.0-beta.7](https://github.com/wont-org/biz-ui/compare/v1.1.0-beta.6...v1.1.0-beta.7) (2025-03-26)
+
+### Bug Fixes
+
+- **EditableSelect:** optionLabelProp ([d5212eb](https://github.com/wont-org/biz-ui/commit/d5212ebad4875c7101746378dbc8b4a91d0dd83b))
+
+# [1.1.0-beta.6](https://github.com/wont-org/biz-ui/compare/v1.1.0-beta.5...v1.1.0-beta.6) (2025-03-26)
+
+### Bug Fixes
+
+- **EditableSelect:** 增加检索功能 ([55b4679](https://github.com/wont-org/biz-ui/commit/55b4679bf61374a76febda26fbffa562660a91c2))
+- **EditableSelect:** disabled ([37bcd5c](https://github.com/wont-org/biz-ui/commit/37bcd5c3f83062a2b4b9936b738fb20733af9708))
+
+### Features
+
+- **EditableSelect:** 增加 afterDelete ([497d77a](https://github.com/wont-org/biz-ui/commit/497d77a2c8c152e9d387dfb40acfd90250d0548d))
+
+# [1.1.0-beta.5](https://github.com/wont-org/biz-ui/compare/v1.1.0-beta.4...v1.1.0-beta.5) (2025-03-24)
+
+### Bug Fixes
+
+- **EditableSelect:** 单选, 删除选项时,值相同则清空 ([e1d789a](https://github.com/wont-org/biz-ui/commit/e1d789a141f2a6e18c71de5019dc50c614ef7871))
+
+# [1.1.0-beta.4](https://github.com/wont-org/biz-ui/compare/v1.1.0-beta.3...v1.1.0-beta.4) (2025-03-24)
+
+### Features
+
+- **EditableSelect:** 删除确认参数透传 ([835a6ba](https://github.com/wont-org/biz-ui/commit/835a6ba6ae827aba0d4c82eeb2c540f5169df581))
+
 # [1.1.0-beta.3](https://github.com/wont-org/biz-ui/compare/v1.1.0-beta.2...v1.1.0-beta.3) (2025-03-10)
 
 ### Features